move ReadBuffer to io, as it is more general than IPC
