+++ src/io/ipc/read/read_basic.rs
@@ -3,12 +3,11 @@ use std::{collections::VecDeque, convert::TryInto};
 
 use crate::buffer::Buffer;
 use crate::error::{Error, Result};
-use crate::io::ipc::read::common::ReadBuffer;
 use crate::{bitmap::Bitmap, types::NativeType};
 
 use super::super::compression;
 use super::super::endianess::is_native_little_endian;
-use super::{Compression, IpcBuffer, Node, OutOfSpecKind};
+use super::{Compression, IpcBuffer, Node, OutOfSpecKind, ReadBuffer};
 
 fn read_swapped<T: NativeType, R: Read + Seek>(
     reader: &mut R,
